disable prompt color by default; add color.sh config
This change requires that the 'color.sh' config value be
enabled to get prompt color:

  $ git config color.sh auto

The --get-color arg used to determine the ANSI colors for the
prompt is apparently fairly new so users with older versions
of git were receiving a warning each time the prompt was
displayed. Older versions of git will not be able to use color
in the prompt but they won't get the annoying warning either.

Thanks to Xavier Shay (http://github.com/xaviershay/git-sh)
for reporting the issue.