Add compatibility for system mv and rm commands.
After glob expansion, each file parameter is tested to see if it is tracked or not. Tracked files are passed to the corresponding git command while, untracked files are passed to the system command (using 'command mv' or 'command rm' as appropriate). Command options (prefixed by '-') are passed through to both git and system commands for all file arguments.